+++ references/example-greeting/src/server/core/repository.go
@@ -1,0 +1,10 @@
+package core
+
+import (
+	skeled "example.com/greeting/skeled/golang"
+)
+
+type GreetingRepository interface {
+	Create(g skeled.Greeting) skeled.Greeting
+	List(offset, limit int) []skeled.Greeting
+}